PSY-425: move collection + request services into community sub-package
Pure relocation: `services/collection.go` + `services/request.go` (with
their tests and compile-time interface assertions) move under a new
`services/community/` sub-package, mirroring the per-domain layout used
by `catalog/`, `engagement/`, `admin/`, etc. No behaviour change.

- New: `services/community/{collection,request}.go` (+ tests) and
  `services/community/interfaces.go` for the
  `CollectionServiceInterface` / `RequestServiceInterface` checks.
- `services/container.go` wires `community.NewCollectionService(db)` /
  `community.NewRequestService(db)`.
- `internal/api/handlers/shared/testhelpers/integration_deps.go` swaps
  to the community sub-package.
- Root `services/interfaces.go` deleted (assertions moved). The
  unused root `services/test_helpers_test.go` (only `stringPtr`, no
  callers in any file that stays at the root) is also removed.
- `internal/api/handlers/community/collection_integration_test.go`
  updates the local `MinPublicCollectionItems` alias to point at
  `community.MinPublicCollectionItems`.
- `docs/runbooks/agent-workflow.md` + `docs/features/collections.md`
  updated to reflect the new path.
