+++ backend/internal/services/community/collection.go
@@ -1,4 +1,4 @@
-package services
+package community
 
 import (
 	"encoding/json"
