+++ backend/internal/services/community/collection_markdown_test.go
@@ -1,4 +1,4 @@
-package services
+package community
 
 import (
 	"strings"
