+++ backend/internal/services/community/interfaces.go
@@ -1,0 +1,9 @@
+package community
+
+import "psychic-homily-backend/internal/services/contracts"
+
+// Compile-time interface satisfaction checks for community services.
+var (
+	_ contracts.CollectionServiceInterface = (*CollectionService)(nil)
+	_ contracts.RequestServiceInterface    = (*RequestService)(nil)
+)
