--- backend/internal/services/test_helpers_test.go
@@ -1,6 +1,0 @@
-package services
-
-// stringPtr returns a pointer to a string. Test helper.
-func stringPtr(s string) *string {
-	return &s
-}
